requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.13.1
sqlalchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
celery[redis]==5.4.0
redis==5.0.8
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0
boto3==1.35.24
python-dateutil==2.9.0.post0
python-multipart==0.0.9
Pillow==10.4.0
numpy==2.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
